fix: refresh project after import
