+++ frontend/src/component/project/Project/Import/import/ImportStage.tsx
@@ -12,8 +12,8 @@ import { PulsingAvatar } from '../PulsingAvatar';
 import { ConditionallyRender } from 'component/common/ConditionallyRender/ConditionallyRender';
 import { Box } from '@mui/system';
 import { useChangeRequestsEnabled } from 'hooks/useChangeRequestsEnabled';
-import useProject from 'hooks/api/getters/useProject/useProject';
 import { usePendingChangeRequests } from 'hooks/api/getters/usePendingChangeRequests/usePendingChangeRequests';
+import { useProjectFeatureSearch } from '../../PaginatedProjectFeatureToggles/useProjectFeatureSearch';
 
 export const ImportStatusArea = styled(Box)(({ theme }) => ({
     padding: theme.spacing(4, 2, 2, 2),
@@ -65,7 +65,7 @@ export const ImportStage: FC<{
     onClose: () => void;
 }> = ({ environment, project, payload, onClose }) => {
     const { createImport, loading, errors } = useImportApi();
-    const { refetch: refreshProject } = useProject(project);
+    const { refetch: refreshProject } = useProjectFeatureSearch(project);
     const { refetch: refreshChangeRequests } =
         usePendingChangeRequests(project);
     const { setToastData } = useToast();
